digit_web_app/requirements.txt
flask==2.0.1
tensorflow==2.6.0
pillow==8.3.1
numpy==1.21.2